+++ dangerzone/global_common.py
@@ -18,8 +18,13 @@ class GlobalCommon(object):
 
     def __init__(self):
         # Version
-        with open(self.get_resource_path("version.txt")) as f:
-            self.version = f.read().strip()
+        try:
+            with open(self.get_resource_path("version.txt")) as f:
+                self.version = f.read().strip()
+        except FileNotFoundError:
+            # In dev mode, in Windows, get_resource_path doesn't work properly for dangerzone-container, but luckily
+            # it doesn't need to know the version
+            self.version = "unknown"
 
         # Initialize terminal colors
         colorama.init(autoreset=True)
@@ -379,7 +384,7 @@ class GlobalCommon(object):
         if self.custom_container:
             return self.custom_container
         else:
-            return "flmcode/dangerzone"
+            return "docker.io/flmcode/dangerzone"
 
     def get_resource_path(self, filename):
         if getattr(sys, "dangerzone_dev", False):
@@ -409,7 +414,7 @@ class GlobalCommon(object):
     def get_dangerzone_container_path(self):
         if getattr(sys, "dangerzone_dev", False):
             # Look for resources directory relative to python file
-            return os.path.join(
+            path = os.path.join(
                 os.path.dirname(
                     os.path.dirname(
                         os.path.abspath(inspect.getfile(inspect.currentframe()))
@@ -418,6 +423,9 @@ class GlobalCommon(object):
                 "dev_scripts",
                 "dangerzone-container",
             )
+            if platform.system() == "Windows":
+                path = f"{path}.bat"
+            return path
         else:
             if platform.system() == "Darwin":
                 return os.path.join(
@@ -431,18 +439,11 @@ class GlobalCommon(object):
                 return "/usr/bin/dangerzone-container"
 
     def exec_dangerzone_container(self, args):
-        # Prefix the args with the retainer runtime, and in the case linux when the user isn't in the docker group, pkexec
-        if platform.system() == "Linux":
-            if self.settings.get("linux_prefers_typing_password"):
-                args = ["/usr/bin/pkexec", self.dz_container_path] + args
-            else:
-                args = [self.dz_container_path] + args
-        else:
-            args = [self.dz_container_path] + args
+        args = [self.dz_container_path] + args
+        args_str = " ".join(pipes.quote(s) for s in args)
+        print(Style.DIM + "> " + Style.NORMAL + Fore.CYAN + args_str)
 
         # Execute dangerzone-container
-        args_str = " ".join(pipes.quote(s) for s in args)
-        print(Fore.YELLOW + "\u2023 " + Fore.CYAN + args_str)  # ‣
         return subprocess.Popen(
             args,
             startupinfo=self.get_subprocess_startupinfo(),
@@ -469,7 +470,7 @@ class GlobalCommon(object):
         ) as p:
             stdout_data, _ = p.communicate()
             lines = stdout_data.split(b"\n")
-            if b"\u2023 " in lines[0]:  # ‣
+            if b"> " in lines[0]:
                 stdout_data = b"\n".join(lines[1:])
 
             # The user canceled, or permission denied