+++ install/linux/build_rpm.py
@@ -6,17 +6,15 @@ import inspect
 import subprocess
 import shutil
 
-sys.path.insert(
-    0, os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-)
-import dangerzone
 
-version = dangerzone.dangerzone_version
 root = os.path.dirname(
     os.path.dirname(
         os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
     )
 )
+
+with open(os.path.join(root, "share", "version.txt")) as f:
+    version = f.read().strip()
 
 
 def main():
@@ -31,7 +29,7 @@ def main():
 
     print("* Building RPM package")
     subprocess.run(
-        "python3 setup.py bdist_rpm --requires='python3-pyside2,python3-appdirs,python3-click,python3-pyxdg,python3-requests,python3-colorama,(docker or docker-ce)'",
+        "python3 setup.py bdist_rpm --requires='podman,python3-pyside2,python3-appdirs,python3-click,python3-pyxdg,python3-requests,python3-colorama'",
         shell=True,
         cwd=root,
         check=True,